Add files via upload
By default, Magnum enters multicast IPs into the 2022-6 IP Inputs of 570ITXEs because it is unaware of 2022-2 sources (all stream types are defined as "Video"). This Reflex document detects the route change specifically into the three available 570ITXE destinations, copies the multicasts across from IP Input to Decoder, then sets the UDX Video IO to accept the Decoder as the input. It then verifies the changes are without discrepancies and runs the process again to correct any detected discrepancies accordingly.